trivial: LayeredMap: break large file
